Move EnsureZKVerifierInitialized to keeper.go
